tests/wpt/tests/html/semantics/scripting-1/the-script-element/css-module-assertions/charset.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CSS modules: UTF-8 decoding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script type="module" onerror="unreachable()">
  import styleSheet from "../serve-with-content-type.py?fn=css-module-assertions/resources/utf-8.css&ct=text/css%3Bcharset=utf-8" assert { type: "css"};
  test(() => {
    assert_equals(styleSheet.rules[0].style.content, "\"śćążź\"");
  }, "CSS module should be loaded as utf-8 when charset=utf8 is specified");
</script>
<script type="module" onerror="unreachable()">
  import styleSheet from "../serve-with-content-type.py?fn=css-module-assertions/resources/utf-8.css&ct=text/css%3Bcharset=shift-jis" assert { type: "css"};
  test(() => {
    assert_equals(styleSheet.rules[0].style.content, "\"śćążź\"");
  }, "CSS module should be loaded as utf-8 when charset=shift-jis is specified");
</script>
<script type="module" onerror="unreachable()">
  import styleSheet from "../serve-with-content-type.py?fn=css-module-assertions/resources/utf-8.css&ct=text/css%3Bcharset=windows-1252" assert { type: "css"};
  test(() => {
    assert_equals(styleSheet.rules[0].style.content, "\"śćążź\"");
  }, "CSS module should be loaded as utf-8 when charset=windows-1252 is specified");
</script>
<script type="module" onerror="unreachable()">
  import styleSheet from "../serve-with-content-type.py?fn=css-module-assertions/resources/utf-8.css&ct=text/css%3Bcharset=utf-7" assert { type: "css"};;
  test(() => {
    assert_equals(styleSheet.rules[0].style.content, "\"śćążź\"");
  }, "CSS module should be loaded as utf-8 when charset=utf-7 is specified");
</script>
<script type="module" onerror="unreachable()">
  import styleSheet from "../serve-with-content-type.py?fn=css-module-assertions/resources/windows-1250.css&ct=text/css%3Bcharset=windows-1250" assert { type: "css"};
  test(() => {
    assert_not_equals(styleSheet.rules[0].style.content, "\"śćążź\"",
                    'Should be decoded as UTF-8');
  }, "CSS module should be loaded as utf-8 even if it is encoded in windows-1250 and served with a windows-1250 charset response header");
</script>
